requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
transformers
torch
Pillow
python-dotenv
pyttsx3
pytest
